// circabc-frontend/src/app/core/generated/circabc/model/shareIGsAndPermissions.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.4.3
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NameValue } from './nameValue';


export interface ShareIGsAndPermissions { 
    igs?: Array<NameValue>;
    permissions?: Array<string>;
}